+++ STM32F4_PBL/Src/DataManagement.cpp
@@ -104,7 +104,7 @@ uint16_t DataManagement::getCollectedData(uint8_t* globalBuffer)
 	std::unique_ptr<uint8_t[]> dataBuffer = std::make_unique<uint8_t[]>(maxDataSize);
 	DataPtrVolumePair* tempObj;
 	uint16_t dataSize = 0;
-	uint8_t returnedSize = 0;
+	uint16_t returnedSize = 0;
 
 	for(auto indexID : refDataArray)
 	{
